feat: claim order by employee

+++ src/main/kotlin/com/rapidex/rapidex_mobile_api/repositories/EmployeeRepository.kt
@@ -3,4 +3,4 @@ package com.rapidex.rapidex_mobile_api.repositories
 import com.rapidex.rapidex_mobile_api.entities.Employee
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface EmployeeRepository : JpaRepository<Employee, Long>
+interface EmployeeRepository : JpaRepository<Employee, Int>
